+++ include/shaders.h
@@ -124,11 +124,11 @@ public:
 
 		// Load to memory
 		shaderFile.seekg(0, ios::end);
-		int shaderFileLength = shaderFile.tellg();			// Get estimation of file size (allways big enough?)
+		int shaderFileLength = (int) shaderFile.tellg();	// Get estimation of file size (allways big enough?)
 		shaderFile.seekg(0, ios::beg);
 		char *shaderSource = new char[shaderFileLength+1];	// Reserve enough memory for the whole file
 		shaderFile.read(shaderSource, shaderFileLength);
-		shaderFileLength = shaderFile.gcount();				// Get real file length
+		shaderFileLength = (int) shaderFile.gcount();		// Get real file length
 		shaderSource[shaderFileLength] = 0;					// Null terminator
 
 		char const *shaderSourceConst = shaderSource;
